+++ src/components/AnswerModal.tsx
@@ -6,8 +6,8 @@ import { PostFamilyAnswer } from '../services/FamilyAnswerApi'
 const AnswerModalBackground = styled.div<{ isOpen: boolean }>`
   display: ${({ isOpen }) => (isOpen ? 'block' : 'none')};
   position: fixed;
-  width: 1440px;
-  height: 1024px;
+  width: 100%;
+  height: 100%;
   left: 0px;
   top: 0px;
 
@@ -20,7 +20,7 @@ const AnswerModalBox = styled.div`
   width: 812px;
   height: 485px;
   left: calc(50% - 812px / 2);
-  top: calc(50% - 485px / 2 + 0.5px);
+  top: 270px;
 
   background: #ffffff;
   border-radius: 34px;
@@ -215,7 +215,7 @@ function AnswerModal({
       alert('답변이 성공적으로 제출되었습니다.')
       toggleModal()
       clearSelectedQuestion()
-      await fetchQuestions(0, 10)
+      await fetchQuestions(0, 45)
     } catch (error) {
       alert('답변 제출 중 오류가 발생했습니다.')
     }
